requirements.txt
amqp==5.2.0
anyio==4.2.0
asgiref==3.7.2
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==4.2.10
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-timezone-field==6.1.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
kombu==5.3.5
packaging==23.2
prompt-toolkit==3.0.43
python-crontab==3.0.0
python-dateutil==2.8.2
python-telegram-bot==20.8
pytz==2024.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
vaderSentiment==3.3.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
